# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management API",
        "description": "API for managing translations with multi-locale and tagging support.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get paginated list of translations",
                "operationId": "ae0f6dd8197765098e77e2b606751d4f",
                "responses": {
                    "200": {
                        "description": "List of translations"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a new translation",
                "operationId": "8ec4f1d3dbc71b9f2731025a94d54165",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "key",
                                    "locale",
                                    "content"
                                ],
                                "properties": {
                                    "key": {
                                        "type": "string"
                                    },
                                    "locale": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "tag": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/translations/{id}": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get a specific translation by ID",
                "operationId": "7bbe51dfbe42fd33f06ebcf0a2845245",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the translation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "key": {
                                            "type": "string"
                                        },
                                        "locale": {
                                            "type": "string"
                                        },
                                        "content": {
                                            "type": "string"
                                        },
                                        "tag": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update an existing translation",
                "operationId": "be40fa3d8222e7f7a35784783720de29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the translation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    },
                                    "tag": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Translation updated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Translations"
                ],
                "summary": "Delete a translation by ID",
                "operationId": "3befb2f8f0f70febf898f90e3ecd33e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the translation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation deleted"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/translations/search": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Search translations by key, content, or tag",
                "operationId": "3e7bf3a699ee7396a6322d1f54bfb0f9",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/translations/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export translations by locale as JSON",
                "operationId": "59c3468414324a9198833b52b8e3bf9b",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exported translations"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Translations",
            "description": "Operations on translations"
        }
    ]
}